Create a servlet utility file
